Calculate and propagate delta tick for scaled ticks
This is to accommodate slower or inconsistent machines. We calculate what is essentially the number of ticks we need to calculate ahead. We also need to move UI related logic OUT of the main game loop (aka out of ticks) and put them into areas they will calculated consistently, to keep the UI fluid.
- Calculate delta tick in Game
- Propagate delta tick to Level and Entity
- Add velocity properties and update logic to the viewport
- Update the viewport in the Renderer

+++ src/main/raptor/engine/nav/agent/DefaultNavAgent.java
@@ -107,7 +107,7 @@ public class DefaultNavAgent implements INavAgent {
 	}
 
 	@Override
-	public void move(final int unitsToMove) {
+	public void move(final double unitsToMove) {
 		if ((path.isEmpty() && atWaypoint()) || unitsToMove < 1)
 			return;
